Delete Course
Created the delete function of the course.

Implemented the JS delete and edit buttons

+++ includes/unit-inc.php
@@ -2,15 +2,30 @@
     require_once "dbh.php";
     require_once "functions.php";
 
-if(isset($_POST["action"])){
-    if($_POST["action"] == "delete");
+if(isset($_GET["action"]) && $_GET["action"]=="list")
+    {
+         $units = getUnits($conn);
+    } 
+
+
+
+if(isset($_POST["action"]) && $_POST["action"] == "delete"){
        $unitId = $_POST['id'];
        deleteUnit($conn, $unitId);
-
-         header("location: ../list-units.php");   
+        header("location: ../list-units.php");   
         exit();
+    }
 
-}
+  if(isset($_POST["action"]) && $_POST["action"] == "edit"){
+    $unit = getUnit($conn, $_POST["id"]);
+
+    $unitName = $unit ['unitName'];
+    $unitDescription =  $unit['unitDescription'];
+    $courseId = $unit['courseId'];
+    $semester = $unit['semester']; 
+  }
+    
+
 
 if (isset($_POST['submit'])) {  
     $unitName = $_POST['unitName'];
@@ -19,7 +34,7 @@ if (isset($_POST['submit'])) {
     $semester = $_POST['semester'];
     addUnit($conn, $unitName, $unitDescription, $courseId, $semester);
 
-      header("location: ../unit.php");   
+      header("location: ../unit.php?success=true");   
         exit();
     
 }
